+++ objectified-ui/public/WHATS_NEW.md
@@ -65,7 +65,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Added simple automatic layout algorithm for arranging nodes (needs further refinement)
   - Added edge styling:
     - Styling for direct, optional, weak, and bidirectional relationships
-    - 
+    - Color selection for edges
 - Added property templates
   - Fixed property copying to a class that contains an object with nested properties
 - Dashboard improvements